32-bit Unit Test Warnings
Some new problems exposed with latest Php8.1.

+++ src/PhpSpreadsheet/Calculation/Engineering/BitWise.php
@@ -17,10 +17,12 @@ class BitWise
      * Split a number into upper and lower portions for full 32-bit support.
      *
      * @param float|int $number
+     *
+     * @return int[]
      */
     private static function splitNumber($number): array
     {
-        return [floor($number / self::SPLIT_DIVISOR), fmod($number, self::SPLIT_DIVISOR)];
+        return [(int) floor($number / self::SPLIT_DIVISOR), (int) fmod($number, self::SPLIT_DIVISOR)];
     }
 
     /**
